+++ madbfs/include/madbfs/data/adb_connection.hpp
