+++ madbfs/src/data/adb_connection.cpp
